+++ src/Symfony/Component/ErrorRenderer/Exception/ErrorRendererNotFoundException.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\ErrorCatcher\Exception;
+namespace Symfony\Component\ErrorRenderer\Exception;
 
 class ErrorRendererNotFoundException extends \RuntimeException
 {