+++ src/Symfony/Component/ErrorRenderer/Exception/FlattenException.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\ErrorCatcher\Exception;
+namespace Symfony\Component\ErrorRenderer\Exception;
 
 use Symfony\Component\Debug\Exception\FatalThrowableError;
 use Symfony\Component\HttpFoundation\Exception\RequestExceptionInterface;